Attempt to fix the submodule_regression test
A number of submodules are currently failing the submodule_regression
test with "NO PRIMARY BRANCH TO TEST!" messages -- this is due to not
having the proper origin/main branch to test against.

I think this can be fixed by doing a `git fetch` in the submodule after
we run a `git submodule sync`, just to make sure the info for the
current URL is up-to-date.